move marlin kernels
